fix(magazine): show the curated edition in the blog magazine modal too
#169 only wired the standalone /magazine page, but the primary entry is
/blog?view=magazine, which renders <MagazineModal> with the raw Bridge feed —
so the curated edition never showed there. Extract the curated fetch into a
shared useCuratedMagazine() hook and use it in BOTH the /magazine page and
MagazineModal (community view): prefer the published edition, fall back to the
live community feed. Profile magazines (explicit posts / hiveUsername) skip it.
